backend/requirements.txt
fastapi
uvicorn
openai
faiss-cpu
pydantic
python-dotenv
sentence-transformers
ollama
pytesseract
python-multipart
pillow
deep-translator
sqlalchemy
python-jose[cryptography]
passlib[bcrypt]
google-auth
